Update computerMarketplace.sol
There is a potential bug in the buyProduct function. The transferFrom function of the IERC20Token interface is called to transfer the Celo tokens from the buyer to the seller. However, the allowance of the buyer is not checked before the transfer is made, which could lead to the buyer accidentally transferring more tokens than intended.

To fix this, it is recommended to add a check to ensure that the buyer has approved the transfer of the exact amount of tokens needed for the purchase. One way to do this is to use the allowance function of the IERC20Token interface to check the approved amount before making the transfer